refactor(yaml): consolidate `yaml/schema/*.ts` code into `yaml/_schema.ts`
