--- yaml/schema/extended.ts
@@ -1,36 +1,0 @@
-// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-// This module is browser compatible.
-
-import { Schema } from "../_schema.ts";
-import { regexp, undefinedType } from "../_type/mod.ts";
-import { DEFAULT_SCHEMA } from "./default.ts";
-
-/***
- * Extends JS-YAML default schema with additional JavaScript types
- * It is not described in the YAML specification.
- * Functions are no longer supported for security reasons.
- *
- * @example
- * ```ts
- * import { parse } from "@std/yaml";
- *
- * const data = parse(
- *   `
- *   regexp:
- *     simple: !!js/regexp foobar
- *     modifiers: !!js/regexp /foobar/mi
- *   undefined: !!js/undefined ~
- * # Disabled, see: https://github.com/denoland/deno_std/pull/1275
- * #  function: !!js/function >
- * #    function foobar() {
- * #      return 'hello world!';
- * #    }
- * `,
- *   { schema: "extended" },
- * );
- * ```
- */
-export const EXTENDED_SCHEMA: Schema = new Schema({
-  explicit: [regexp, undefinedType],
-  include: [DEFAULT_SCHEMA],
-});
